+++ src/components/pages/Mypage/styles/index.js
@@ -1,8 +1,8 @@
 import StyledMypage from './Mypage.styles';
-import StyledMyPageButton from './MypageButton.styles';
+import StyledMypageButton from './MypageButton.styles';
 import StyledMyArticlePreview from './MyArticlePreview.styles';
 
 import StyledLinkBox from './Linkbox.styles';
 import StyledMyProfile from './MyProfle.styles';
 
-export { StyledMypage, StyledMyPageButton, StyledMyArticlePreview, StyledLinkBox, StyledMyProfile };
+export { StyledMypage, StyledMypageButton, StyledMyArticlePreview, StyledLinkBox, StyledMyProfile };
